[Changed] Factored out saving new item and location to respective DAO classes [1.19.0]
