+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.19.0-SNAPSHOT] - 2020-10-04
+### Changed
+- Factored out saving new item and location to respective DAO classes
+
 ## [1.18.0-SNAPSHOT] - 2020-10-04
 ### Added
 - Generate a monthly report pdf with Jasper reports
