+++ src/main/java/com/zeylin/pricetracker/dao/ItemDAO.java
@@ -8,6 +8,7 @@ import org.jooq.DSLContext;
 import org.jooq.Record3;
 import org.jooq.Result;
 import org.jooq.SelectJoinStep;
+import org.jooq.impl.DSL;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
@@ -33,5 +34,25 @@ public class ItemDAO {
 
         Result<Record3<Integer, String, Integer>> records = from.orderBy(i.ITEM_ID).fetch();
         return records.map(r -> PriceConverter.convertToItemDto(i, r));
+    }
+
+    /**
+     * Save an item.
+     * @param itemName item name
+     * @return ID of the saved item
+     */
+    public Integer saveItem(String itemName) {
+        Item i = Item.ITEM;
+
+        Integer itemId = dsl.select(DSL.max(i.ITEM_ID).plus(1))
+                .from(i)
+                .fetchOne()
+                .value1();
+
+        dsl.insertInto(i)
+                .set(i.ITEM_ID, itemId)
+                .set(i.NAME, itemName)
+                .execute();
+        return itemId;
     }
 }
